feat(db) add CommodityAuction EntityType - each instantiation will be 1 row in the db - this entity type is part of a list on the CommodityAuctionSnapshot entity type, has a many to 1 relationship with it - this type allows me to store each individual auction as a row in the DB

+++ wow-paper-trader.Ingestor/Persistence/EntityTypes/CommodityAuctionSnapshot.cs
@@ -2,12 +2,13 @@
 
 public sealed class CommodityAuctionSnapshot
 {
-    public int Id { get; private set; }
+    public long Id { get; private set; }
 
-    public int IngestionRunId { get; private set; }
+    public long IngestionRunId { get; private set; }
 
     //this is a navigation property, it is not created as a column in the db
     //it is for linking this entity type to the ingestionRun in code
+    //one to one relationship
     public IngestionRun IngestionRun { get; private set; } = null!;
 
     public DateTime FetchedAtUtc { get; private set; }
@@ -15,9 +16,10 @@ public sealed class CommodityAuctionSnapshot
     public string ApiEndPoint { get; private set; } = string.Empty;
 
     //navigation collection property becuase this is a list of custom types
+    //one to many relationship
     public List<CommodityAuction> CommodityAuctions { get; private set; } = new List<CommodityAuction>();
 
-    public CommodityAuctionSnapshot(int ingestionRunId, DateTime fetchedAtUtc, string apiEndPoint)
+    public CommodityAuctionSnapshot(long ingestionRunId, DateTime fetchedAtUtc, string apiEndPoint)
     {
         ingestionRunId = IngestionRunId;
         fetchedAtUtc = FetchedAtUtc;
